[#1517] convert L1 memory functions to use void* instead of unsigned long for addresses
git-svn-id: svn://localhost/svn/uclinux-dist/trunk@3341 3e2a5524-39e8-4cb9-84a9-c49da69cba3a

+++ linux-2.6.x/drivers/video/bfin_ad7171fb_main.c
@@ -697,9 +697,9 @@ static int bfin_ad7171_fb_open(struct fb_info *info, int user)
 static int bfin_ad7171_fb_release(struct fb_info *info, int user)
 {
 	if (rgb_l1)
-		l1_data_A_sram_free((unsigned long)rgb_l1);
+		l1_data_A_sram_free(rgb_l1);
 	if (yuv_l1)
-		l1_data_A_sram_free((unsigned long)yuv_l1);
+		l1_data_A_sram_free(yuv_l1);
 	del_timer(&bfin_framebuffer_timer);
 	bfin_disable_dma();
 	bfin_disable_ppi();
